split MRSolver.hs into multiple files
